# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Options trading platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stock {
  id: number;
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  volume: number;
  marketCap: number;
  sector: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  technicalStrength: number;
}

export type StockDetailLiquidity =
  (typeof StockDetailLiquidity)[keyof typeof StockDetailLiquidity];

export const StockDetailLiquidity = {
  Very_Liquid: "Very Liquid",
  Liquid: "Liquid",
  Moderate: "Moderate",
  Low: "Low",
} as const;

export interface StockDetail {
  id: number;
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  volume: number;
  marketCap: number;
  sector: string;
  technicalStrength: number;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  eps: number;
  pe: number;
  dividendYield: number;
  ivRank: number;
  relativeStrength: string;
  supportPrice: number;
  resistancePrice: number;
  earningsDate: string;
  liquidity: StockDetailLiquidity;
  priceAction: string;
}

export interface PricePoint {
  date: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface WatchlistItem {
  id: number;
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  technicalStrength: number;
  addedAt: string;
}

export interface AddToWatchlistBody {
  symbol: string;
}

export type OptionsStrategyType =
  (typeof OptionsStrategyType)[keyof typeof OptionsStrategyType];

export const OptionsStrategyType = {
  trade: "trade",
  income: "income",
} as const;

export type OptionsStrategyOutlook =
  (typeof OptionsStrategyOutlook)[keyof typeof OptionsStrategyOutlook];

export const OptionsStrategyOutlook = {
  bullish: "bullish",
  bearish: "bearish",
  neutral: "neutral",
} as const;

export type OptionLegAction =
  (typeof OptionLegAction)[keyof typeof OptionLegAction];

export const OptionLegAction = {
  buy: "buy",
  sell: "sell",
} as const;

export type OptionLegOptionType =
  (typeof OptionLegOptionType)[keyof typeof OptionLegOptionType];

export const OptionLegOptionType = {
  call: "call",
  put: "put",
} as const;

export interface OptionLeg {
  action: OptionLegAction;
  optionType: OptionLegOptionType;
  strikePrice: number;
  premium: number;
  quantity: number;
  expiration: string;
}

export interface OptionsStrategy {
  id: number;
  name: string;
  type: OptionsStrategyType;
  outlook: OptionsStrategyOutlook;
  description: string;
  legs: OptionLeg[];
  tradeCost: number;
  maxProfit: number;
  maxLoss: number;
  returnPercent: number;
  breakeven: number;
  /**
   * @minimum 0
   * @maximum 200
   */
  score: number;
  expirationDate: string;
}

export interface PnlRequest {
  strategyId: number;
  targetPrice: number;
  targetDate: string;
  impliedVolatility: number;
}

export type PnlResultPnlCurveItem = {
  price: number;
  pnl: number;
};

export interface PnlResult {
  profitLoss: number;
  profitLossPercent: number;
  breakeven: number;
  maxProfit: number;
  maxLoss: number;
  pnlCurve: PnlResultPnlCurveItem[];
}

export type DashboardSummaryMarketSentiment =
  (typeof DashboardSummaryMarketSentiment)[keyof typeof DashboardSummaryMarketSentiment];

export const DashboardSummaryMarketSentiment = {
  bullish: "bullish",
  bearish: "bearish",
  neutral: "neutral",
} as const;

export interface DashboardSummary {
  totalStocks: number;
  watchlistCount: number;
  avgTechnicalStrength: number;
  bullishCount: number;
  bearishCount: number;
  neutralCount: number;
  marketSentiment: DashboardSummaryMarketSentiment;
}

export interface TopMovers {
  gainers: Stock[];
  losers: Stock[];
}

export type ListStocksParams = {
  search?: string;
  limit?: number;
};

export type GetStockPriceHistoryParams = {
  period?: GetStockPriceHistoryPeriod;
};

export type GetStockPriceHistoryPeriod =
  (typeof GetStockPriceHistoryPeriod)[keyof typeof GetStockPriceHistoryPeriod];

export const GetStockPriceHistoryPeriod = {
  "1D": "1D",
  "1W": "1W",
  "1M": "1M",
  "3M": "3M",
  "6M": "6M",
  "1Y": "1Y",
} as const;

export type GetStrategiesParams = {
  outlook?: GetStrategiesOutlook;
};

export type GetStrategiesOutlook =
  (typeof GetStrategiesOutlook)[keyof typeof GetStrategiesOutlook];

export const GetStrategiesOutlook = {
  bullish: "bullish",
  bearish: "bearish",
  neutral: "neutral",
} as const;