# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Options trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryMarketSentiment } from "./dashboardSummaryMarketSentiment";

export interface DashboardSummary {
  totalStocks: number;
  watchlistCount: number;
  avgTechnicalStrength: number;
  bullishCount: number;
  bearishCount: number;
  neutralCount: number;
  marketSentiment: DashboardSummaryMarketSentiment;
}