# lib/api-zod/src/generated/types/optionLeg.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Options trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OptionLegAction } from "./optionLegAction";
import type { OptionLegOptionType } from "./optionLegOptionType";

export interface OptionLeg {
  action: OptionLegAction;
  optionType: OptionLegOptionType;
  strikePrice: number;
  premium: number;
  quantity: number;
  expiration: string;
}