# lib/api-zod/src/generated/types/optionsStrategy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Options trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OptionLeg } from "./optionLeg";
import type { OptionsStrategyOutlook } from "./optionsStrategyOutlook";
import type { OptionsStrategyType } from "./optionsStrategyType";

export interface OptionsStrategy {
  id: number;
  name: string;
  type: OptionsStrategyType;
  outlook: OptionsStrategyOutlook;
  description: string;
  legs: OptionLeg[];
  tradeCost: number;
  maxProfit: number;
  maxLoss: number;
  returnPercent: number;
  breakeven: number;
  /**
   * @minimum 0
   * @maximum 200
   */
  score: number;
  expirationDate: string;
}