# lib/api-zod/src/generated/types/pnlRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Options trading platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PnlRequest {
  strategyId: number;
  targetPrice: number;
  targetDate: string;
  impliedVolatility: number;
}