# lib/api-zod/src/generated/types/pnlResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Options trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PnlResultPnlCurveItem } from "./pnlResultPnlCurveItem";

export interface PnlResult {
  profitLoss: number;
  profitLossPercent: number;
  breakeven: number;
  maxProfit: number;
  maxLoss: number;
  pnlCurve: PnlResultPnlCurveItem[];
}