# lib/api-zod/src/generated/types/topMovers.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Options trading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Stock } from "./stock";

export interface TopMovers {
  gainers: Stock[];
  losers: Stock[];
}